requirements.txt
asgiref==3.3.4
Django==3.2.4
django-js-asset==1.2.2
django-mptt==0.13.3
django-mptt-admin==2.1.0
pkg-resources==0.0.0
psycopg2==2.9.1
pytz==2021.1
russian-names==0.1.2
sqlparse==0.4.2
typing-extensions==3.10.0.0
